Updated tetris.py
Added double move left and right with Z and X. Fixed known issues 1 and 2 from last version, primarily that the end screen not displaying when a block has overfilled the screen. Divided some logic into their own functions. Removed an unnecessary sort for the block tuples. Added a disclaimer on the start for the minimum dimension to avoid curses error.